requirements.txt
torch==2.0.1
timm==0.9.5
gradio==3.41.0
matplotlib==3.7.2
numpy==1.24.4
open3d==0.17.0
Pillow==10.0.0
torchvision==0.15.2